pyproject.toml
[project]
name = "webagent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
